pyproject.toml
[tool.poetry]
name = "demo-api-test"
version = "0.1.0"
description = ""
authors = ["Andrey.Ch <zona2033@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
allure-pytest = "^2.13.2"
py-zabbix = "^1.1.7"
beautifulsoup4 = "^4.12.2"
pytest-rerunfailures = "^12.0"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
